Docs: Make AJAX search field also work when pressing enter
When a user presses enter, the same AJAX query is triggered
as when entering some letters. As a result, the form is
never really submitted and no new page appears.
However, if for some reason AJAX does not work (in particular, if the
onSubmit action handler isn't called), then a search page is loaded
with the same results as the AJAX query...