Fix 763 Fretboards don't transpose properly
A transpose that took a chord out of the default octave broke the
default fretboards.

The code is now modified to look both an octave up and an octave down
from the given octave if no entry is found for the given octave.
(cherry picked from commit 9bf6dcaf6ca27f7141ec825bb3569de8a5090dbb)