Fix vertical placement of first system.
The top system on a page is now positioned relative to the
bottom of the top-margin (ie. the top of the header) instead
of the bottom of the header as it was before. This allows
the top line to be placed more consistently if the
header varies from page to page.