texinfo/SConstruct: Use SConsic rather than pythonic constructs.
Also, move texiBuilder to builder.py.  This makes for a much more
fair comparison with make -- apart from the fact that examples/
and images/ contain duplicate file names, both of which are
attempted to be installed in out/.  SCons does not like this,
so we need some ugly coding until this is resolved.