texinfo/manuals.itexi: Short menu texts for manuals.  Fixes too wide tocs.
And esp. the toc for Manuals is much too wide.  Possibly we have too
many entries in "Manuals".

For this to work, apply the included patch

    texinfo/texi2html-toc-shortest-of-name-or-text.patch

[After merging]: Unfortunately, while making this patch, the number
of toc entries for Manuals went up from 11 to 12 :-)

TODO: figure out what text to make short, the node name or the text.

    * NODE :: TEXT

this texi2html patch just takes the shortest non-empty one as TOC entry :-)